requirements.txt
# Core LangChain and LangGraph
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.38
langgraph>=0.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.0.0

# Web Framework
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
